pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cson-forge"
version = "0.1.0"
description = "A utility for generating regional oceanographic modeling domains and spawning reproducible C-Star workflows"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "C-Worthy Ocean"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "jinja2",
    "pandas",
    "pyyaml>=6.0",
    "roms-tools>=3.3.0",
    "copernicusmarine",
    "gdown",
]

[tool.setuptools]
# Only include the cson_forge package
# tests and workflows directories are at root level and automatically excluded
packages = ["cson_forge"]

[tool.setuptools.package-data]
"*" = ["*.yml", "*.yaml"]

